+++ src/services/privateDns/format.ts
@@ -25,8 +25,21 @@ export default ({
     numberOfVirtualNetworkLinksWithRegistration,
     provisioningState,
     internalId,
-    resourceGroupId
+    resourceGroupId,
+    records = [],
+    virtualNetworkLinks = []
   } = service
+
+  // Records
+  const aRecords = records.filter(r => r.type === 'A').map(r => r.id)
+  const aaaaRecords = records.filter(r => r.type === 'AAAA').map(r => r.id)
+  const mxRecords = records.filter(r => r.type === 'MX').map(r => r.id)
+  const ptrRecords = records.filter(r => r.type === 'PTR').map(r => r.id)
+  const soaRecord = records.filter(r => r.type === 'SOA').map(r => r.id)
+  const srvRecords = records.filter(r => r.type === 'SRV').map(r => r.id)
+  const txtRecords = records.filter(r => r.type === 'TXT').map(r => r.id)
+  const cnameRecord = records.filter(r => r.type === 'CNAME').pop()?.id
+
   return {
     id,
     subscriptionId: account,
@@ -43,6 +56,15 @@ export default ({
     provisioningState,
     internalId,
     tags: formatTagsFromMap(Tags),
-    resourceGroupId
+    resourceGroupId,
+    aRecords,
+    aaaaRecords,
+    mxRecords,
+    ptrRecords,
+    soaRecord,
+    srvRecords,
+    txtRecords,
+    cnameRecord,
+    virtualNetworkLinks: virtualNetworkLinks.map(r => r.id)
   }
 }